cookbook/ingredients/serializers.py
from rest_framework import serializers

from cookbook.ingredients.models import Category, Ingredient


class CategorySerializer(serializers.HyperlinkedModelSerializer):
    class Meta:
        model = Category
        fields = ['name']


class IngredientSerializer(serializers.HyperlinkedModelSerializer):
    class Meta:
        model = Ingredient
        fields = ['name', 'notes', 'category']
